src/requirements.txt
amba-event-stream
requests
pymongo
lxml